% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enough_comps.R
\name{enough_comps}
\alias{enough_comps}
\title{Find out if parcel has enough comparables}
\usage{
enough_comps(dat, ID, mile_radius = 3, n_sale = 5)
}
\arguments{
\item{dat}{data frame containing lat, long, and sales information of all parcels}

\item{ID}{character or numeric of serial number of parcel you want to test}

\item{n_sale}{numeric for number of sales required to have enough comparables}

\item{miles_radius}{numeric for number of miles you want your range to be to search for sales}
}
\value{
something
}
\description{
This function returns whether a parcel has enough comparables to do a sales grid to get an estimated value for a parcel
Best used with purrr::map to apply over all parcels
}
\examples{
ID <- "094837839"
dat <- lat_long_data
value_from_cap_rate(dat,ID, mile_radius = 1, n_sale =10)
}
