% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_cap_rate.R
\name{find_cap_rate}
\alias{find_cap_rate}
\title{Find cap rate}
\usage{
find_cap_rate(net_income, sale_price)
}
\arguments{
\item{net_income}{Numeric vector of net income}

\item{sale_price}{Numeric vector of actual sale prices}
}
\value{
A named list of estimated cap rate, including upper and lower bounds:
mean_estimate, mean_lowerbound, and mean_upperbound
}
\description{
This function returns a named list of estimated cap rate based on income and sale values from similar property types.
It can handle missing, NaN, and Inf values.
}
\examples{
net_income <- c(51329.93,76826.74,42861.05,34066.37,142085.22,81460.92)
sale_price <- c(NA,1600000,0,1050000,1050000,479000)
cap_rate <- find_cap_rate(net_income,sale_price)
cap_rate$mean_estimate; cap_rate$mean_lowerbound; cap_rate$mean_upperbound
}
