% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouped_ratio_stats.R
\name{grouped_ratio_stats}
\alias{grouped_ratio_stats}
\title{Build a grouped ratio analysis comparable to NCSS output.}
\usage{
grouped_ratio_stats(
  df,
  assessed_column,
  sold_price_column,
  grouping_column,
  output = "list",
  min_count = 10
)
}
\arguments{
\item{df}{Data frame containing columns with assessed values, sold prices, and a grouping variable at the minimum}

\item{assessed_column}{The name (not quoted) of the column with assessed values}

\item{sold_price_column}{The name (not quoted) of the column with the sale prices you wish to compare}

\item{grouping_column}{The name (not quoted) of the column containing grouping data (e.g., neighborhood)}

\item{output}{A character vector that must be either "list" (default) or "table" that determines how the output is presented. "list" returns a list of lists. "table" returns something similar to the NCSS output.}

\item{min_count}{Integer (default=10) determining the minimum number of observations. Cannot be less than 3.}
}
\value{
Either a list or table of ratio statistics, broken down by a grouping variable.
}
\description{
This function returns either a named list of Ratio Statistics comparing predicted and actual
sale prices, OR it can return a table comparable to NCSS
}
\details{
This function requires a grouping variable, such as a Neighborhood
It will not perform tests for any groups that have fewer than 3 observations, but will still display them if output="table"

The NormTest_Results field is to match how NCSS decides whether to recommend Mean or Median. It is based on a Shapiro test of normality with a 0.10 cutoff.
}
\examples{
grouped_ratio_stats(df,
  assessed_column = factor_value,
  sold_price_column = time_adjusted_sold_price,
  grouping_column = nbhd,
  output="table")
}
