% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MedianCI.R
\name{MedianCI}
\alias{MedianCI}
\title{Generate median with estimated confidence intervals}
\usage{
MedianCI(
  x,
  conf.level = 0.95,
  sides = c("two.sided", "left", "right"),
  na.rm = FALSE,
  method = c("exact", "boot"),
  R = 999
)
}
\arguments{
\item{x}{a (non-empty) numeric vector of data values}

\item{conf.level}{confidence level of the interval}

\item{sides}{a character string specifying the side of the confidence interval, must be one of "two.sided" (default), "left" or "right". You can specify just the initial letter. "left" would be analogue to a hypothesis of "greater" in a t.test.}

\item{na.rm}{logical. Defaults to FALSE}

\item{method}{defining the type of interval that should be calculated (one out of "exact", "boot"). Default is "exact".}

\item{R}{The number of bootstrap replicates. Usually this will be a single positive integer.}
}
\value{
A numeric vector with 3 elements: median, lwr.ci, upr.ci
}
\description{
Taken from DescTools::StripAttr
}
\examples{
MedianCI(rnorm(100),na.rm=TRUE)
}
