% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_districts.R
\name{merge_districts}
\alias{merge_districts}
\title{Merge districts}
\usage{
merge_districts(
  df,
  N = 50,
  district_id_col = "district",
  sold_price_col = "sold_price",
  res_rcnld_col = "res_rcnld",
  rl_value_col = "rl_value",
  acreage_col = "acreage"
)
}
\arguments{
\item{df}{Data frame containing, at a minimum: district, rcnld, and sold_price. These can be named anything and specified in the function.}

\item{district_id_col}{Character string specifying the name of the column that specifies district. Default = "district"}

\item{sold_price_col}{Character string specifying the name of the column containing sale price values. Default = "sold_price"}

\item{res_rcnld_col}{Character string specifying the name of the column containing rcnld or other measure for comparing districts. Default = "res_rcnld"}

\item{rl_value_col}{Character string specifying the name of the column containing land_value or other measure for comparing districts. Default = "rl_value"}

\item{acreage_col}{Character string specifying the name of the column containing acreage or other measure for comparing districts. Default = "acreage"}
}
\value{
A named list of estimated cap rate, including upper and lower bounds:
mean_estimate, mean_lowerbound, and mean_upperbound
}
\description{
This function finds disctricts with low sales volume and merges them into similar larger districts.
Similarity is determined as the "high-volume" district with the closest median res_rcnld (or other column).
The function returns an updated data frame with a new column "new_district"
}
\examples{
df <- data("property_info")
merged <- merge_districts(df)
merged \%>\%
pivot_longer(c(district,new_district),
             names_to = "type",
             values_to = "district") \%>\%
             mutate(district=factor(district),
             type=case_when(type == "district" ~ "old district",
                            TRUE ~ "new district")) \%>\%
ggplot(aes(x=district,fill=type)) + geom_bar()
}
