% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_multiple_years.R
\name{pivot_multiple_years}
\alias{pivot_multiple_years}
\title{Pivot multiple years}
\usage{
pivot_multiple_years(dat, years)
}
\arguments{
\item{dat}{Data frame containing multiple years of valuation data spread out over columns. This assumes that colnames are clean (i.e., x2022_VAR1,x2022_VAR2) and start with the year number.}

\item{years}{integer or numeric vector for the years to pull and pivot. These values are converted to character class and used to search in regex}
}
\value{
A transformed data set, reduced to the listed years, and converted to long format.
}
\description{
This function helps when dealing with pivoting multiple years of assessment values. Data is commonly found with columns like:
2022_re_res, 2022_re_agr, 2022_re_com, 2021_re_res, etc.  This "wide format" can be cleaned with pivot_longer, but this function
provides a way of dealing with multiple years. It splits the data frame into pieces by year, and then joins them together in long format.
}
\examples{
long <- pivot_multiple_years(df,2018:2020)

}
