% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratio_stats.R
\name{ratio_stats}
\alias{ratio_stats}
\title{Build a ratio analysis comparable to SPSS "Ratio Statistics".}
\usage{
ratio_stats(assessed_value, sold_prices, CI = 0.95)
}
\arguments{
\item{assessed_value}{Numeric vector of assessed values (must be same length as sold_prices)}

\item{sold_prices}{Numeric vector of actual sold prices (must be same length as preds)}

\item{CI}{The confidence interval between 0 and 1. Defaults to .95}
}
\value{
A named list of ratio statistics:
Mean (incl. lower and upper bounds), Median (incl. lower and upper bounds),
Actual coverage, Price related differential, Coefficient of dispersion, and
median centered coefficient of variation. The "Ratio" refers to Assessed Value / Sold Price
}
\description{
This function returns a named list of Ratio Statistics comparing predicted and actual
sale prices
}
\examples{
preds <- runif(100,min = 100000,500000)
sales <- runif(100,min = 100000,500000)
RatioAnalysis(preds = preds,sold_prices = sales, CI = .95)
When Price_Related_Differential less than .98, high value properties are over appraised
When Price_Related_Differential greater than 1.03, high value properties are under appraised
}
