% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ratio_Analysis.R
\name{RatioAnalysis}
\alias{RatioAnalysis}
\title{Build a ratio analysis for a trained real estate valuation model.}
\usage{
RatioAnalysis(dat = dat, model = mod, pred = "pred", CI = 0.95)
}
\arguments{
\item{dat}{A data frame of property sales data that was used to train a model.
This is the 'raw' data.}

\item{model}{A trained model object}

\item{pred}{The name of the column of predicted values as a character string. Defaults to "pred"}

\item{CI}{The confidence interval between 0 and 1. Defaults to .95}
}
\value{
A data frame of ratio statistics for the model:
Mean (incl. lower and upper bounds), Median (incl. lower and upper bounds),
Actual coverage, Price related differential, Coefficient of dispersion, and
median centered coefficient of variation. The "Ratio" refers to Assessed Value / Sold Price
}
\description{
This function will auto-determine the name of the outcome variable based on your model formula.
It takes, as inputs, the raw property data, the trained model, and returns a custom ratio analysis.
This serves as a replacement for SPSS 'Ratio Analysis'
}
\examples{
df <- data.frame(BR=c(1,2,3,4,1,2,3,4,1,2,3,4),SQFT=c(1000,1500,1750,2500,1100,1250,1550,2350,1350,1600,2000,2300),ASSESSED=c(300000,350000,475000,560000,349000,387000,421000,622000,368000,402000,521000,600000),#'    SOLD=c(295000,356000,470000,555000,342000,380000,411000,630000,363000,407000,501000,610000))
mod <- glm(SOLD ~ BR + SQFT,data=df)
RatioAnalysis(dat=df,model = mod)
}
