% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StripAttr.R
\name{StripAttr}
\alias{StripAttr}
\title{Remove attributes from an object}
\usage{
StripAttr(x, attr_names = NULL)
}
\arguments{
\item{x}{the object whose attributes should be removed}

\item{attr_names}{a vector with attribute names, which will be removed. Leaving the default to NULL will cause all the attributes to be deleted.}
}
\value{
the object x without the attributes contained in attr_names
}
\description{
For convenience we sometimes want to strip some or all attributes in a oneliner.
Taken from DescTools::StripAttr
}
\examples{
x <- runif(10)
StripAttr(x)
}
