% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value_from_cap_rate.R
\name{value_from_cap_rate}
\alias{value_from_cap_rate}
\title{Estimate value from income and cap rate}
\usage{
value_from_cap_rate(net_income, cap_rate)
}
\arguments{
\item{net_income}{Numeric vector of net income}

\item{cap_rate}{Numeric vector of length 1 indicating the cap rate for similar properties}
}
\value{
something
}
\description{
This function returns a vector of estimated property values based on net income and cap rate
It can handle missing, NaN, and Inf values.
}
\examples{
net_income <- c(51329.93,76826.74,42861.05,34066.37,142085.22,0,NA,Inf)
cap_rate <- 0.0964
value_from_cap_rate(net_income,cap_rate)
}
